+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 * Implemented `RemAssign<Hex>` for `Hex`
 * Implemented `Rem<i32>` for `Hex`
 * Implemented `RemAssign<i32>` for `Hex`
+* Added angle methods to `Direction`:
+  - `angle_flat` for radian angle in flat orientation
+  - `angle_flat_degrees` for degrees angle in flat orientation
+  - `angle_pointy` for radian angle in pointy orientation
+  - `angle_pointy_degrees` for degrees angle in pointy orientation
+  - `angle` for radian angle in a given orientation
+* (**BREAKING**) rotated order of `Direction` enum and `Hex::ALL_NEIGHBORS` by 1 to the left, `TopRight` is now the first as `BottomRight` is now last
 
 ## 0.2.0